requirements.txt
aiohttp==3.9.5
boto3==1.34.130
cryptography==42.0.8
discord.py @ git+https://github.com/Rapptz/discord.py.git
Django==5.0.4
numpy==1.26.4
psycopg==3.1.18
psycopg-pool==3.2.2
PyNaCl==1.5.0
python-decouple==3.8
python-dotenv==1.0.1
torch==2.3.1